Moved linking vinyls/items to later time at runtime - this should ensure the target item is properly initialised
