+++ FP2Lib/Stage/StageHandler.cs
@@ -91,8 +91,6 @@ namespace FP2Lib.Stage
             if (!Stages.ContainsKey(stage.uid))
             {
                 stage.id = AssignStageID(stage, stage.isHUB);
-                stage.itemID = GetStageItem(stage);
-                stage.vinylID = GetStageMusic(stage);
                 stage.registered = true;
                 Stages.Add(stage.uid, stage);
                 WriteToStorage();
@@ -104,8 +102,6 @@ namespace FP2Lib.Stage
                 stage.id = Stages[stage.uid].id;
                 Stages[stage.uid] = stage;
                 Stages[stage.uid].id = AssignStageID(stage, stage.isHUB);
-                Stages[stage.uid].itemID = GetStageItem(stage);
-                Stages[stage.uid].vinylID = GetStageMusic(stage);
                 Stages[stage.uid].registered = true;
                 WriteToStorage();
                 return true;
@@ -136,32 +132,6 @@ namespace FP2Lib.Stage
                 if (stage.id == id && stage.isHUB) return stage;
             }
             return null;
-        }
-
-        private static FPMusicTrack GetStageMusic(CustomStage stage)
-        {
-            if (!stage.vinylUID.IsNullOrWhiteSpace())
-            {
-                VinylData vinyl = VinylHandler.GetVinylDataByUid(stage.vinylUID);
-                if (vinyl != null)
-                {
-                    return (FPMusicTrack)vinyl.id;
-                }
-            }
-            return stage.vinylID;
-        }
-
-        private static FPPowerup GetStageItem(CustomStage stage)
-        {
-            if (!stage.itemUID.IsNullOrWhiteSpace())
-            {
-                //TODO: Implement code once ItemHandler is done
-                return FPPowerup.NONE;
-            } 
-            else
-            {
-                return stage.itemID;
-            }
         }
 
         private static int AssignStageID(CustomStage stage, bool isHub)
